chore: re-structure file hierarchy
