+++ src/business/dto/transactionDto.ts
@@ -3,7 +3,7 @@ import { Either } from "../../framework/shared/either"
 import { IError } from "../../framework/shared/iError"
 
 export interface InputCreateTransactionDto {
-  bankAccountId: string
+  bankAccountId?: string
   userId: string
   name: string
   valueCents: Number
