+++ src/business/useCases/categories/createCategoryUseCase.ts
@@ -1,10 +1,10 @@
 import { injectable, inject } from 'inversify'
 
-import { left, right } from '../../framework/shared/either'
-import { IUseCase } from './iUseCase'
-import { InputCreateCategoryDto, OutputCreateCategoryDto } from '../dto/categoryDto'
-import { CategoryCreationFailed } from '../module/errors/categories'
-import { ICategoryRepository, ICategoryRepositoryToken } from '../repositories/iCategoryRepository'
+import { left, right } from '../../../framework/shared/either'
+import { IUseCase } from '../iUseCase'
+import { InputCreateCategoryDto, OutputCreateCategoryDto } from '../../dto/categoryDto'
+import { CategoryCreationFailed } from '../../module/errors/categories'
+import { ICategoryRepository, ICategoryRepositoryToken } from '../../repositories/iCategoryRepository'
 
 @injectable()
 export class CreateCategoryUseCase implements IUseCase<InputCreateCategoryDto, OutputCreateCategoryDto> {
