+++ src/business/useCases/categories/deleteCategoryUseCase.ts
@@ -1,10 +1,10 @@
 import { injectable, inject } from 'inversify'
 
-import { left, right } from '../../framework/shared/either'
-import { IUseCase } from './iUseCase'
-import { InputDeleteCategoryDto, OutputDeleteCategoryDto } from '../dto/categoryDto'
-import { CategoryDeleteFailed, CategoryNotFound } from '../module/errors/categories'
-import { ICategoryRepository, ICategoryRepositoryToken } from '../repositories/iCategoryRepository'
+import { left, right } from '../../../framework/shared/either'
+import { IUseCase } from '../iUseCase'
+import { InputDeleteCategoryDto, OutputDeleteCategoryDto } from '../../dto/categoryDto'
+import { CategoryDeleteFailed, CategoryNotFound } from '../../module/errors/categories'
+import { ICategoryRepository, ICategoryRepositoryToken } from '../../repositories/iCategoryRepository'
 
 @injectable()
 export class DeleteCategoryUseCase implements IUseCase<InputDeleteCategoryDto, OutputDeleteCategoryDto> {
