+++ src/controller/serializers/categories/inputCreateCategory.ts
@@ -1,9 +1,9 @@
 import { IsHexColor, IsMongoId, IsNotEmpty, IsString } from 'class-validator'
 
-import { ICategoryEntity } from '../../domain/entities/categoryEntity'
-import { Either } from '../../framework/shared/either'
-import { IError } from '../../framework/shared/iError'
-import { Validatable } from './abstractValidatable'
+import { ICategoryEntity } from '../../../domain/entities/categoryEntity'
+import { Either } from '../../../framework/shared/either'
+import { IError } from '../../../framework/shared/iError'
+import { Validatable } from '../abstractValidatable'
 
 export class InputCreateCategory extends Validatable<InputCreateCategory> {
   @IsNotEmpty()
