+++ src/controller/serializers/transactions/inputGetTransaction.ts
@@ -1,9 +1,9 @@
 import { IsMongoId, IsNotEmpty } from 'class-validator'
-import { ITransactionEntity } from '../../domain/entities/transactionEntity'
 
-import { Either } from '../../framework/shared/either'
-import { IError } from '../../framework/shared/iError'
-import { Validatable } from './abstractValidatable'
+import { ITransactionEntity } from '../../../domain/entities/transactionEntity'
+import { Either } from '../../../framework/shared/either'
+import { IError } from '../../../framework/shared/iError'
+import { Validatable } from '../abstractValidatable'
 
 export class InputGetTransaction extends Validatable<InputGetTransaction> {
   @IsNotEmpty()
