+++ src/framework/functions/categories/createCategory.ts
@@ -5,8 +5,8 @@ import { APIGatewayProxyEvent, Context } from 'aws-lambda'
 import { httpHandler } from '../../utility/httpHandler'
 import { container } from '../../shared/ioc/container'
 import { httpResponse } from '../../utility/httpResponse'
-import { CreateCategoryOperator } from '../../../controller/operators/createCategoryOperator'
-import { InputCreateCategory } from '../../../controller/serializers/inputCreateCategory'
+import { CreateCategoryOperator } from '../../../controller/operators/categories/createCategoryOperator'
+import { InputCreateCategory } from '../../../controller/serializers/categories/inputCreateCategory'
 
 export const handler = httpHandler(async (event: APIGatewayProxyEvent, context: Context) => {
   context.callbackWaitsForEmptyEventLoop = false
