+++ src/framework/functions/transactions/deleteTransaction.ts
@@ -5,8 +5,8 @@ import { APIGatewayProxyEvent, Context } from 'aws-lambda'
 import { httpHandler } from '../../utility/httpHandler'
 import { container } from '../../shared/ioc/container'
 import { httpResponse } from '../../utility/httpResponse'
-import { DeleteTransactionOperator } from '../../../controller/operators/deleteTransactionOperator'
-import { InputDeleteTransaction } from '../../../controller/serializers/inputDeleteTransaction'
+import { DeleteTransactionOperator } from '../../../controller/operators/transactions/deleteTransactionOperator'
+import { InputDeleteTransaction } from '../../../controller/serializers/transactions/inputDeleteTransaction'
 
 export const handler = httpHandler(async (event: APIGatewayProxyEvent, context: Context) => {
   context.callbackWaitsForEmptyEventLoop = false
