+++ src/framework/functions/transactions/getTransaction.ts
@@ -5,8 +5,8 @@ import { APIGatewayProxyEvent, Context } from 'aws-lambda'
 import { httpHandler } from '../../utility/httpHandler'
 import { container } from '../../shared/ioc/container'
 import { httpResponse } from '../../utility/httpResponse'
-import { GetTransactionOperator } from '../../../controller/operators/getTransactionOperator'
-import { InputGetTransaction } from '../../../controller/serializers/inputGetTransaction'
+import { GetTransactionOperator } from '../../../controller/operators/transactions/getTransactionOperator'
+import { InputGetTransaction } from '../../../controller/serializers/transactions/inputGetTransaction'
 import { TransactionNotFound } from '../../../business/module/errors/transactions'
 
 export const handler = httpHandler(async (event: APIGatewayProxyEvent, context: Context) => {
