+++ src/framework/ioc/operatorModule.ts
@@ -1,9 +1,9 @@
 import { ContainerModule, interfaces } from 'inversify'
-import { CreateCategoryOperator } from '../../controller/operators/createCategoryOperator'
-import { CreateTransactionOperator } from '../../controller/operators/createTransactionOperator'
-import { DeleteCategoryOperator } from '../../controller/operators/deleteCategoryOperator'
-import { DeleteTransactionOperator } from '../../controller/operators/deleteTransactionOperator'
-import { GetTransactionOperator } from '../../controller/operators/getTransactionOperator'
+import { CreateCategoryOperator } from '../../controller/operators/categories/createCategoryOperator'
+import { DeleteCategoryOperator } from '../../controller/operators/categories/deleteCategoryOperator'
+import { CreateTransactionOperator } from '../../controller/operators/transactions/createTransactionOperator'
+import { DeleteTransactionOperator } from '../../controller/operators/transactions/deleteTransactionOperator'
+import { GetTransactionOperator } from '../../controller/operators/transactions/getTransactionOperator'
 
 export const OperatorModule = new ContainerModule((bind: interfaces.Bind) => {
   bind(CreateTransactionOperator).toSelf()
